Gut the repository
I plan to use only the eneyj/data/ part of this fork while migrating
GraalEneyj to the Wikifunctions model (e.g. I will *not* be updating the
eneyj JS code accordingly), and I don’t want anybody to mistake this
repository for something that it’s not, so let’s remove almost
everything else.