Migrate top-level objects to Z2/persistent object
In the Wikifunctions model, the actual stored objects should all have
the type Z2, and the “proper” value is stored below the key Z2K2. The
labels also move to a key of Z2, and descriptions/aliases/documentation
are discarded for now (there’s no equivalent for them at the moment).
The whole migration is implemented in jq, mostly because it would take
ages to do it by hand even for just the objects needed by GraalEneyj,
but that should also be useful in case we need to redo the migration on
slightly modified input later for some reason.