Replace language objects with language codes
In the AbstractText model, the language of a monolingual text object is
a reference to a language object; in the WikiLambda extension, it’s a
plain language code string so far – though in the Wikifunctions model as
documented, it’s still a reference. So maybe this isn’t final yet, but
it should be easy enough to undo if needed.