Migrate implementation layout
This migrates the layout *within* implementation objects. It does *not*
yet migrate how the implementations are attached to the function. (I’m
not even sure what that part is supposed to look like in the new model,
to be honest.) The fixed Z14K1 key, pointing to an “implementation
implementation” object of varying type, goes away; instead, a function
call is attached under Z14K2, a code object under Z14K3, or a builtin is
marked by Z14K4. (In the new model, the value of Z14K4 can be an ID of
the builtin, but in the old model builtins had no IDs so we just create
strings for them out of thin air.) Once that migration is done, the ID
of the function being implemented is attached as the new Z14K1 (builtins
already had this under Z1K2).

When encountering any unknown implementation type (Z8, Z18, Z200, or
sometimes the Z14K1 was even an array instead of an object), stash it
away under Z14K5 (unassigned in the new function model so far), to be
dealt with later.